::sparkles: feat: adiciona filtro por DY e inclui dados de fundamentos
